Expand configuration query
Expand configuration query to be able to return multiple configurations.
Change name to configurations. Make beamline parameter optional (and
rename to beamlineFilter). When a filter is provided, only details for
that beamline are returned, otherwise all beamline configurations are
returned.

Also add a test for the multi-configuration DB query and update the
example query in the README.